NFC: rename device folder to cable_type
