+++ src/deluge/gui/menu_item/midi/devices.cpp
@@ -20,9 +20,9 @@
 #include "gui/ui/sound_editor.h"
 #include "hid/display/display.h"
 #include "io/debug/log.h"
-#include "io/midi/devices/din.h"
-#include "io/midi/devices/usb_device_cable.h"
-#include "io/midi/devices/usb_hosted.h"
+#include "io/midi/cable_types/din.h"
+#include "io/midi/cable_types/usb_device_cable.h"
+#include "io/midi/cable_types/usb_hosted.h"
 #include "io/midi/midi_device.h"
 #include "io/midi/midi_device_manager.h"
 #include "util/container/static_vector.hpp"
