+++ src/deluge/io/midi/cable_types/din.h
