+++ src/deluge/io/midi/cable_types/usb_common.cpp
