+++ src/deluge/io/midi/cable_types/usb_device_cable.cpp
