+++ src/deluge/io/midi/cable_types/usb_device_cable.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "deluge/io/midi/devices/usb_common.h"
+#include "deluge/io/midi/cable_types/usb_common.h"
 
 class MIDIDeviceUSBUpstream final : public MIDICableUSB {
 public:
