+++ src/deluge/io/midi/midi_device_manager.cpp
@@ -20,9 +20,9 @@
 #include "gui/menu_item/mpe/zone_num_member_channels.h"
 #include "gui/ui/sound_editor.h"
 #include "hid/display/display.h"
+#include "io/midi/cable_types/din.h"
+#include "io/midi/cable_types/usb_device_cable.h"
 #include "io/midi/device_specific/specific_midi_device.h"
-#include "io/midi/devices/din.h"
-#include "io/midi/devices/usb_device_cable.h"
 #include "io/midi/midi_device.h"
 #include "io/midi/midi_engine.h"
 #include "mem_functions.h"
